pyproject.toml
[project]
name = "guardrails"
version = "1.0.0"
description = "A pluggable image and text guardrails pipeline for validating and sanitizing content before AI processing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "python-magic>=0.4.27; sys_platform != 'win32'",
    "certifi>=2023.0.0",
    "opennsfw2>=0.10.0",
    "tensorflow>=2.13.0",
    "opencv-python>=4.8.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "presidio-image-redactor>=0.0.48",
    "pytesseract>=0.3.10",
    "imagehash>=4.3.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
    "python-multipart>=0.0.22",
]

[project.scripts]
guardrails = "guardrails.cli:main"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/guardrails"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
